requirements.txt
pytest==8.3.4
allure-pytest==2.13.5
pytest-playwright==0.7.0
pydantic==2.10.6
pydantic-settings==2.7.1
pillow==11.2.1

